Add files via upload
Project created as part of an assignment for a Seneca College course called CIA710 Professional Communication of Analysis. Data from Toronto Police Service open data portal was downloaded and analyzed to create recommendations in a presentation. I used JN to create a folium map in Python 3 with embedded visualization as part of cluster markers popups with their associated coordinates.